+++ packages/core/src/lib.js
@@ -4,6 +4,7 @@ export * from './lib/date';
 export * from './lib/debounce';
 export * from './lib/dom';
 export * from './lib/events';
+export * from './lib/options';
 export * from './lib/stores';
 export * from './lib/utils';
 export * from './lib/view';