+++ packages/day-grid/src/index.js
@@ -1,11 +1,12 @@
 import {writable} from 'svelte/store';
-import {btnTextMonth, intl} from '@event-calendar/core';
+import {btnTextMonth, intl, themeView} from '@event-calendar/core';
 import {days} from './stores.js';
 import View from './View.svelte';
 
 export default {
 	createOptions(options) {
 		options.dayMaxEvents = false;
+		options.dayCellFormat = {day: 'numeric'};
 		options.dayPopoverFormat = {month: 'long', day: 'numeric', year: 'numeric'};
 		options.moreLinkContent = undefined;
 		// Common options
@@ -13,7 +14,6 @@ export default {
 		options.buttonText.close = 'Close';
 		options.theme.uniform = 'ec-uniform';
 		options.theme.dayFoot = 'ec-day-foot';
-		options.theme.month = 'ec-month';
 		options.theme.popup = 'ec-popup';
 		options.view = 'dayGridMonth';
 		options.views.dayGridMonth = {
@@ -22,12 +22,14 @@ export default {
 			dayHeaderFormat: {weekday: 'short'},
 			displayEventEnd: false,
 			duration: {months: 1},
+			theme: themeView('ec-day-grid ec-month-view'),
 			titleFormat: {year: 'numeric', month: 'long'}
 		};
 	},
 
 	createStores(state) {
 		state._days = days(state);
+		state._intlDayCell = intl(state.locale, state.dayCellFormat);
 		state._intlDayPopover = intl(state.locale, state.dayPopoverFormat);
 		state._hiddenEvents = writable({});
 		state._popupDate = writable(null);