+++ packages/time-grid/src/View.svelte
@@ -6,15 +6,13 @@
     import Day from './Day.svelte';
     import Week from './all-day/Week.svelte';
 
-    let {_viewDates, _intlDayHeader, _viewClass, _scrollable, allDaySlot, theme} = getContext('state');
-
-    $_viewClass = 'week';
+    let {_viewDates, _intlDayHeader, allDaySlot, theme} = getContext('state');
 </script>
 
 <div class="{$theme.header}">
     <Section>
         {#each $_viewDates as date}
-            <div class="{$theme.day}" use:setContent={$_intlDayHeader.format(date)}></div>
+            <div class="{$theme.day} {$theme.weekdays?.[date.getUTCDay()]}" use:setContent={$_intlDayHeader.format(date)}></div>
         {/each}
     </Section>
     <div class="{$theme.hiddenScroll}"></div>